+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [4.0.2] - Unreleased
+### Added
+- Add an option to format settings that allows emulating weekly format in a way
+  where top level sections will be named after weeks. This does not apply to subsections.
 ### Changed
 - Course index always shows all sections and activities regardless of the current page. More details
   https://github.com/marinaglancy/moodle-format_flexsections/issues/39